pyproject.toml
[tool.poetry]
name = "dotfm"
version = "0.1.0"
description = "Personal website"
authors = ["Tobi DEGNON <tobidegnon@proton.me>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
gunicorn = "^20.1.0"
psycopg2 = "2.9.3"
django-fbv = "^0.4.0"
django-environ = "^0.9.0"
django-htmx = "^1.12.1"
whitenoise = "^6.2.0"
django-tailwind = "^3.4.0"
django-model-utils = "^4.2.0"
django-lifecycle = "^1.0.0"
django-taggit = "^3.0.0"
django-extensions = "^3.2.0"
markdown2 = "^2.4.3"
django-q = "^1.3.9"
django-crispy-forms = "^1.14.0"
crispy-tailwind = "^0.5.0"

[tool.poetry.dev-dependencies]
django-browser-reload = "^1.6.0"
mypy = "^0.971"
black = "^22.6.0"
isort = "^5.10.1"
pytest = "^7.1.2"
poetryup = "^0.10.0"
pywatchman = "^1.4.1"
django-stubs = "^1.12.0"
pylint-django = "^2.5.3"
pre-commit = "^2.20.0"
ipython = "^8.4.0"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
